refactor: split cli output helpers
